requirements.txt
Flask==3.0.2
flask-cors==4.0.0

tensorflow==2.10.0
numpy==1.23.5
pillow==9.5.0
gunicorn==21.2.0